Bug fix #137 Serial authentification failed
Fixed serial authentification failed problem when AUTOLOGIN on.

This problem is due to the parallel action of login when AUTOLOGIN on.
Applying a series action instead of parallel one can solve the probleme.
The error result object is different from the previous one thus the way
to obtain the error message has been changed accordingly.