Take rack layout into account in 3D view
In 3D view, the relative location of rack was never rendered as defined
in rack.xml, which does not fully represent the real situation. The
factor posy has been introduced to the 3D view instead of merely
counting on the number of rack per row. The laters make every rack from
the same row side by side, even if they are not adjacent.

Fix #151